requirements.txt
Django==5.0.1
djangorestframework==3.14.0
python-decouple==3.8
Pillow>=11.0.0
requests==2.31.0
django-allauth==0.61.1
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn
whitenoise
psycopg2-binary
dj-database-url